Adds model init tests
